backend/agent/pyproject.toml
[project]
name = "healthcare-agent"
version = "0.1.0"
description = "A2UI healthcare discharge form agent powered by Google ADK and Gemini"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["discharge_form_agent"]
